requirements.txt
# Core web framework
Flask>=2.0.0
Flask-WTF>=1.0.0
Werkzeug>=2.0.0

# Image processing and computer vision
Pillow>=9.0.0
opencv-python>=4.5.0
scikit-image>=0.18.0

# Scientific computing and numerical operations
numpy>=1.20.0
scipy>=1.7.0

# Noise generation for effects
noise>=1.2.2

# Environment configuration (optional but recommended)
python-dotenv>=0.19.0

# Production server (for deployment)
gunicorn>=20.1.0
